# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Order APIs",
        "description": "Order API description",
        "version": "1.0.0"
    },
    "paths": {
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get list of orders",
                "description": "Returns list of orders",
                "operationId": "showOrders",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page for order listing",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit of records per page",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "[{\"id\": 1,\"distance\": 15,\"status\": \"unassigned\"}]"
                    },
                    "406": {
                        "description": "Unable to fetch orders"
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create orders",
                "description": "Create orders",
                "operationId": "createOrders",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "origin": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "destination": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "{'id':25,'distance':'634923 m','status':'UNASSIGNED'}"
                    },
                    "406": {
                        "description": "Unable to create orders"
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "patch": {
                "tags": [
                    "Orders"
                ],
                "summary": "Take orders",
                "description": "Take orders",
                "operationId": "takeOrders",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "status": "TAKEN"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "{\"status\":\"SUCCESS\"}"
                    },
                    "406": {
                        "description": "{\"error\":\"Order is already taken\"}"
                    }
                }
            }
        }
    }
}